Completed rotational cipher exercise
This exercise deserves a few notes:

0) My approach was to create a dictionary that maps the original alphabet to the rotated
alphabet.  This was we can quickly translate each letter to its rotated letter.

1) I initially implemented with itertools.cycle and itertools.islice. This allowed
me to create a continguous cycle of a-z + a-z to slice from. While this worked it was
unncessarily complex.

2) Upon further research and reflection, it's clear that we don't need a cycle. In English,
we have 26 letters, so the dictionary must have 26 key/value pairs (x2 for upper case). The
second set of letters can be cut at the rotation key and added to the end. For example, if
rotation key is 3 ('a' -> 'd'), then the keys are 'a-z' and the values are 'd-z' + 'a-c',
which can be coded as ascii_lowercase[key:] + ascii_lowercase[:key].

3) The syntax for creation a dictionary out of two sets of iterables are dict(zip(interable1, iterable2)))

4) The rotator.get(character, character) is a bit strange, but operates exactly as .get() advertises.
It attempts to lookup the character to rotate. If it can't, it uses the original character.