Convert map function to tail-call recursion.
This pattern uses the accumulator pattern to allow the return function to just call the recursive function rather than
and passing in what's accumulating into the function rather than accumulating it outside. Example:

return map(function, tail, accumulator + [function(item)])

instead of

return [function(item)] + map(function, tail)

The latter function builds up a large stack of [function(items)] + [function(items)] + ... + map(function, tail)
in the cases of very large lists and can lead to a stack overflow. Whereas the former calculates the accumulator everytime and
passes it into the next recursion call. The system doesn't need to build up a stack when it's just a "return map(..)".
Tracing the returns back up, we can see that it just returns the final value through the recusion calls. The system recognizes
this and can shortcut the returns and not need to maintain a deep stack, thus avoiding stack overflows.