Completed matching brackets exercise
This is an exericise that ensures all of the brackets in an input text has matching bracket elements.
My approach is to create a stack of opening brackets and iterate through the text. If we see an opening bracket, we append it to the stack.
If we see a closing bracket, we ensure that it matches the last preceeding opening bracket, popped from our stack.  At the end of the text, there should be no
opening brackets left unclosed.